move core model and service provider to lib
