+++ controllers/auth_controller.lua
@@ -4,7 +4,7 @@ local User = require('models.user')
 local AccountLog = require('models.account_log')
 local validator = require('lib.validator')
 local config = require('config.app')
-local auth = require('providers.auth_service_provider')
+local auth = require('lib.auth_service_provider')
 local cjson = require('cjson')
 local user_service = require('services.user_service')
 local random = require('lib.random')
