+++ docker/nginx/conf/nginx.conf
@@ -29,7 +29,7 @@ http {
         server_name localhost;
         
         lua_need_request_body off;
-        lua_code_cache off;
+        lua_code_cache on;
         
         location / {
             content_by_lua_file '/app/bootstrap.lua';
