+++ lib/auth_service_provider.lua
