+++ middleware/token_refresh.lua
@@ -1,4 +1,4 @@
-local auth_service = require('providers.auth_service_provider')
+local auth_service = require('lib.auth_service_provider')
 
 local _M = {}
